// src/contracts/generated/maverick-schema.d.ts
export interface components {
  schemas: {
// ...
    Cart: {
      uuid: string;
      emailAddress: string;
      phoneNumber?: string;
      /** Format: int64 */
      eventId: number;
      /** Format: int64 */
      listingId: number;
      /** Format: int32 */
      quantity: number;
      firstName?: string;
      lastName?: string;
      pricePerTicket: number;
      serviceFee: number;
      deliveryFee: number;
      taxAmount: number;
      availableDeliveryMethods?: components['schemas']['DeliveryMethod'][];
      deliveryMethod: components['schemas']['DeliveryMethod'];
      promoDiscount: number;
      promoDescription?: string;
      giftCardRedemptions?: components['schemas']['GiftCardRedemption'][];
      section: string;
      row: string;
      notes: string;
      seatNumbers?: string;
      event: components['schemas']['Event'];
      emailOptIn: boolean;
      smsOptIn: boolean;
      listing: components['schemas']['Listing'];
      shipToCountryCode?: string;
      /** Format: int64 */
      orderId?: number;
      checkoutVersion?: string;
      experimentId?: string;
      /** Format: int32 */
      variationId?: number;
      amountToCharge?: number;
      serviceFeePerTicket?: number;
      cartTotal?: number;
    };
// ...
    Address: {
      phoneNumber?: string;
      firstName: string;
      lastName: string;
      company?: string;
      postalCode: string;
      address1: string;
      address2?: string;
      city: string;
      state: string;
      country: string;
      save?: boolean;
    };
// ...
    ProductSpecificMessage: {
      type?: string;
      uri?: string;
    };
// ...
    PlaceOrderRequest: {
      shippingAddress?: components['schemas']['Address'];
      billingAddress: components['schemas']['Address'];
      shippingIsBilling?: boolean;
      /** @description Required if the cart has an outstanding balance. */
      paymentNonce?: string;
      /** @enum {string} */
      paymentMethod?: 'applePay' | 'googlePay' | 'payPal' | 'creditCard';
      customer: components['schemas']['Customer'];
      emailAddress: string;
      phoneNumber: string;
      cartUuid: string;
      hasInsurance?: boolean;
      insuranceFee?: number;
      orderTotal: number;
      /** Format: int64 */
      deliveryMethodId: number;
      confirmedMultipleEventOrders?: boolean;
      braintreeDeviceData?: string;
      unitBilled?: number;
      serviceFee?: number;
      deliveryFee?: number;
      taxAmount?: number;
      devicePayload?: string;
      insuranceData?: components['schemas']['InsuranceData'];
      cardVerificationNonce?: string;
      postponedInsuranceOffer?: boolean;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}